+++ service/database.py
@@ -11,13 +11,12 @@ logger = logging.getLogger(__name__)
 # create sqlalchemy object
 db = SQLAlchemy()
 
-# constants
-MAX_COURSE_NUM_LEN: int = 7
+
 
 # ====================================
 # DATABASE MODELS
 # ====================================
-class User(db.Model):
+class Account(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     username = db.Column(db.String(20), nullable=False, unique=True)
     password = db.Column(db.LargeBinary, nullable=False) # hashed password as byte string (don't have to mess with encode/decode)
@@ -60,19 +59,6 @@ class InstructorComment(db.Model):
     instructor_name = db.Column(db.String(20), nullable=False)
     content = db.Column( db.Text, nullable=False )
     timestamp = db.Column( db.DateTime, default=datetime.utcnow)
-
-    def serialize( self ):
-        # get username from id for display
-        username = fetch_username(self.user_id)
-
-        return {
-            'id' : self.id,
-            'user_id': self.user_id,
-            'username': username,
-            'instructor_name': self.instructor_name,
-            'content': self.content,
-            'timestamp': self.timestamp.isoformat()
-        }
     
 class CourseComment(db.Model):
     id = db.Column( db.Integer, primary_key=True, autoincrement = True )
@@ -80,19 +66,6 @@ class CourseComment(db.Model):
     class_name = db.Column(db.String(10), nullable=False)
     content = db.Column( db.Text, nullable=False )
     timestamp = db.Column( db.DateTime, default=datetime.utcnow)
-
-    def serialize( self ):
-        # get username from id for display
-        username = fetch_username(self.user_id)
-
-        return {
-            'id' : self.id,
-            'user_id': self.user_id,
-            'username': username,
-            'class_name': self.class_name,
-            'content': self.content,
-            'timestamp': self.timestamp.isoformat()
-        }
 
 class InstructorRating(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
@@ -151,74 +124,6 @@ def fetch_grade_distribution_data(db: SQLAlchemy):
     db.session.bulk_save_objects(data_to_add)
     db.session.commit()
 
-def add_comment(username, review_type, content):
-    try:
-        # Get user id by provided username
-        user_id = fetch_user_id(username)
-
-        # Create a new Comment object with user_id, instructor/course name, and content
-        if len(review_type) <= MAX_COURSE_NUM_LEN:
-            new_comment = CourseComment(user_id=user_id, class_name=review_type, content=content)
-        else:
-            new_comment = InstructorComment(user_id=user_id, instructor_name=review_type, content=content)
-        
-        # Add the new comment to the database session
-        db.session.add(new_comment)
-        
-        # Commit the session to save changes
-        db.session.commit()
-        
-        return new_comment  # Return the newly created comment
-    except Exception as database_error:
-        # Roll back the session in case of error
-        db.session.rollback()
-        
-        return None
-
-def fetch_comments(search_by):
-    if len(search_by) <= MAX_COURSE_NUM_LEN:
-        return CourseComment.query.filter_by(class_name=search_by).all()
-
-    return InstructorComment.query.filter_by(instructor_name=search_by).all()
-
-# def delete_comment(comment):
-#     # Check if the comment exists
-#     db.session.delete(comment)
-#     db.session.commit()
-#     return True
-    
-def fetch_classes(class_name: str):
-    # get all classes (id, name) from database that match the string up to that point
-    return db.session.query(ClassData).with_entities(ClassData.class_nbr, ClassData.class_name).filter_by(class_name=class_name).all()
-
-def search_for(search: str):
-        
-    # strip any non-alphanumeric characters
-    search = ''.join(e for e in search if e.isalnum())
-
-    # find the first digit in the string
-    numIndex = 0
-    for index, char in enumerate(search):
-        if char.isdigit():
-            numIndex = index
-            break
-
-    # if there is a number, search for class name
-    if numIndex:
-
-        # search for class name 
-        search_results = ClassData.query.with_entities(ClassData.class_name).filter(ClassData.class_name.ilike(f"%{search[:numIndex]} {search[numIndex:]}%")).distinct().all()
-        search_results = [name[0] for name in search_results]
-        return [search_results, "class"]
-
-    # assume name and make them distinct
-    instructor_names = ClassData.query.with_entities(ClassData.instructor_name).filter(ClassData.instructor_name.ilike(f"%{search}%")).distinct().all()
-
-    # combine the two lists into search_results
-    search_results = [name[0] for name in instructor_names]
-
-    return [search_results, "instructor"]
-
 def add_rating(user_id, search_name, rating, by):
     try:
         # Check if user has already made a rating for this instructor/class
@@ -262,16 +167,4 @@ def rating_exists(user_id, search_name, by):
         # return if the user has already given a rating to this instructor
         return InstructorRating.query.filter_by(user_id=user_id, instructor_name=search_name).first()
     
-
-def fetch_user_id(username):
-    # get user from database using username
-    user = User.query.filter_by(username=username).first()
-    # return user id
-    return user.id
-
-
-def fetch_username(user_id):
-    user = User.query.filter_by(id=user_id).first()
-
-    return user.username
   # ====================================
